fix: переместить функционал Movies в App, поправить преобразрование movie.country в строку
