+++ src/components/App/App.js
@@ -11,13 +11,16 @@ import NotFoundPage from '../NotFoundPage/NotFoundPage';
 import Main from '../Main/Main';
 import Footer from '../Footer/Footer';
 import { mainApi } from '../../utils/MainApi';
+import { moviesApi } from '../../utils/MoviesApi';
 import { CurrentUserContext } from '../../contexts/CurrentUserContext';
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
+import moviesFilter from '../../utils/MoviesFilter'
 
 function App() {
 
   const history = useHistory();
   const location = useLocation();
+  const currentViewport = document.documentElement.clientWidth;
 
   const [loggedIn, setLoggedIn] = useState(false);
   const [currentUser, setCurrentUser] = useState({});
@@ -26,11 +29,35 @@ function App() {
   const [loginErrorMessage, setLoginErrorMessage] = useState('');
   const [profileMessage, setProfileMessage] = useState('');
 
+  const [initialMovies, setInitialMovies] = useState([]);
+  const [renderedMovies, setRenderedMovies] = useState([]);
+  const [filteredMovies, setFilteredMovies] = useState([]);
+  const [request, setRequest] = useState('');
+  const [checkboxStatus, setCheckboxStatus] = useState(false);
+  
+  const [preloader, setPreloader] = useState(false);
+  const [searchStatus, setSearchStatus] = useState('');
+  const [isSearchDone, setIsSearchDone] = useState(false);
+    
+  const [firstResults, setFirstResults] = useState(0);
+  const [moreResults, setMoreResults] = useState(0);
+  const [moreButtonVisibility, setMoreButtonVisibility] = useState(false);
+
   const [savedMovies, setSavedMovies] = useState([]);
 
   useEffect(() => {
     tokenCheck();
   }, [loggedIn])
+
+  useEffect(() => {
+    if (localStorage.getItem('moviesStorage')) {
+      const initialSearch = JSON.parse(localStorage.getItem('moviesStorage'));
+      const searchResult = moviesFilter(initialSearch, request, checkboxStatus);
+
+      setFilteredMovies(searchResult);
+      setIsSearchDone(true);
+    }
+  }, [])
 
   useEffect(() => {
     mainApi.getSavedMovies()
@@ -40,6 +67,8 @@ function App() {
       })
       .catch((err) => console.log(err));
   }, [currentUser])
+
+  //Функционал Регистрация/Логин/Профиль
 
   function handleRegisterSubmit(userData) {
     mainApi.register(userData)
@@ -110,6 +139,88 @@ function App() {
       })
   }
 
+  //Функционал Фильмы
+
+  function launchPreloader() {
+    setPreloader(true);
+    setTimeout(() => setPreloader(false), 700);
+  }
+
+  function handleSearchMovie(request, checkboxStatus) {
+    launchPreloader();
+    setRenderedMovies([]);
+    setRequest(request);
+    setCheckboxStatus(checkboxStatus);
+
+    const initialMoviesInLocalStorage = JSON.parse(localStorage.getItem('initialMovies'));
+
+    if (!initialMoviesInLocalStorage) {
+      setPreloader(true);
+      moviesApi.getMovies()
+        .then((moviesData) => {
+          setInitialMovies(moviesData);
+          localStorage.setItem('initialMovies', JSON.stringify(moviesData));
+        })
+        .catch(() => {
+          setSearchStatus('Во время запроса произошла ошибка. Возможно, проблема с соединением или сервер недоступен. Подождите немного и попробуйте ещё раз.')
+        })
+        .finally(() => {
+          setPreloader(false);
+        })
+    } else {
+      setInitialMovies(initialMoviesInLocalStorage);
+    }
+  }
+  
+  useEffect(() => {
+    if (initialMovies.length > 0) {
+      const moviesStorage = moviesFilter(initialMovies, request, checkboxStatus);
+      
+      localStorage.setItem('moviesStorage', JSON.stringify(moviesStorage));
+      localStorage.setItem('request', request);
+      localStorage.setItem('checkboxStatus', checkboxStatus);
+
+      setFilteredMovies(moviesStorage);
+      setIsSearchDone(true);
+    }
+  }, [initialMovies, request, checkboxStatus]);
+
+  //Изменение ширины экрана и показ кол-ва карточек
+
+  useEffect(() => {
+    if (currentViewport <= 480) {
+      setFirstResults(5);
+      setMoreResults(2);
+    } else if (currentViewport <= 768) {
+      setFirstResults(8);
+      setMoreResults(2);
+    } else if (currentViewport > 768) {
+      setFirstResults(12);
+      setMoreResults(8);
+    }
+  }, [currentViewport]);
+  
+  useEffect(() => {
+    if (filteredMovies.length > 0) {
+      if (filteredMovies.length > firstResults) {  
+        setRenderedMovies(filteredMovies.slice(0, firstResults));
+        setMoreButtonVisibility(true);
+      } else {
+        setRenderedMovies(filteredMovies);
+      }
+    }
+  }, [filteredMovies, firstResults]);
+
+  function renderMovies() {
+    setRenderedMovies((state) => filteredMovies.slice(0, state.length + moreResults));
+  }
+
+  useEffect(() => {
+    if (renderedMovies.length === filteredMovies.length) {
+      setMoreButtonVisibility(false);
+    }
+  }, [renderedMovies, filteredMovies]);
+
   function handleSaveMovie(movie) {
     mainApi.saveMovie(movie)
       .then((newMovie) => {
@@ -151,9 +262,16 @@ function App() {
               path="/movies"
               component={Movies}
               loggedIn={loggedIn}
+              onSearch={handleSearchMovie}
+              preloader={preloader}
+              isSearchDone={isSearchDone}
+              searchStatus={searchStatus}
+              renderedMovies={renderedMovies}
               savedMovies={savedMovies}
               onSaveMovie={handleSaveMovie}
               onDeleteMovie={handleDeleteMovie}
+              moreButtonVisibility={moreButtonVisibility}
+              onRenderMovies={renderMovies}
             />
 
             <ProtectedRoute
